Refactoring pipelines part 1
