+++ src/PSRule/Pipeline/InputPathBuilder.cs
@@ -6,7 +6,7 @@ namespace PSRule.Pipeline;
 /// <summary>
 /// A builder for input paths.
 /// </summary>
-internal sealed class InputPathBuilder(IPipelineWriter logger, string basePath, string searchPattern, IPathFilter filter, IPathFilter required)
+internal sealed class InputPathBuilder(IPipelineWriter logger, string basePath, string searchPattern, IPathFilter filter, IPathFilter? required)
     : PathBuilder(logger, basePath, searchPattern, filter, required)
 {
 }
